requirements.txt
# PyTorch core
torch>=2.0.0
torchvision>=0.15.0

# Image processing and augmentation
numpy>=1.21.0
opencv-python>=4.6.0
albumentations>=1.3.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0